+++ generate_successors.cpp
@@ -6,17 +6,18 @@
 #include "get_delta_t.h"
 #include "scheduler.h"
 #include "pruning_constraints.h"
+#include <cinttypes>
 
 
 using namespace std;
 
 
 
-bool is_processor_available_to_tau_i(const state& s, const unsigned short i, const unsigned short m) {
+bool is_processor_available_to_tau_i(const state& s, const uint8_t i, const uint8_t m) {
     
     unsigned short hpJobsNum = 0;
     
-    for (unsigned short indx = 0; indx < i; indx++) {
+    for (uint8_t indx = 0; indx < i; indx++) {
         if (s.c[indx] > 0) {
             hpJobsNum++;
             if (hpJobsNum == m) break;
@@ -35,15 +36,15 @@ bool is_processor_available_to_tau_i(const state& s, const unsigned short i, con
 
 
 // Successors are not sorted
-void generate_successors(const TS& ts, state& s, const unsigned short m, vector<state>* successors, unsigned short& successorsNum) {
+void generate_successors(const TS& ts, state& s, const uint8_t m, vector<state>* successors, unsigned short& successorsNum) {
     
-    int curSuccessorsNum;
+    unsigned short curSuccessorsNum;
     
     // Generate successor states
     (*successors)[0] = s;
     successorsNum = 1;
 
-    for (int i = 0; i < ts.n; i++) {
+    for (uint8_t i = 0; i < ts.n; i++) {
         
         for (unsigned short successorItr = 0; successorItr < successorsNum; successorItr++)
             (*successors)[successorItr].processorAvailableForTau_i[i] = is_processor_available_to_tau_i((*successors)[successorItr], i, m);
@@ -54,7 +55,7 @@ void generate_successors(const TS& ts, state& s, const unsigned short m, vector<
             if ((i == (ts.n-1)) && (!s.prevState_processorAvailableForTau_i[i])) continue;
             
             curSuccessorsNum = successorsNum;
-            for (int successorItr = 0; successorItr < curSuccessorsNum; successorItr++) {
+            for (unsigned short successorItr = 0; successorItr < curSuccessorsNum; successorItr++) {
                 
                 // Davis and Burns 2011 condition B for tau_k and any tau_i
                 if (!condition_cri_2(s, m, i, (*successors)[successorItr].processorAvailableForTau_i[i]))
@@ -73,7 +74,7 @@ void generate_successors(const TS& ts, state& s, const unsigned short m, vector<
     for (unsigned short successorItr = 0; successorItr < successorsNum; successorItr++) {
         (*successors)[successorItr].updateCounters();
         
-        for (unsigned short i = 0; i < ts.n; i++)
+        for (uint8_t i = 0; i < ts.n; i++)
             (*successors)[successorItr].processorAvailableForTau_i[i] = is_processor_available_to_tau_i((*successors)[successorItr], i, m);
     }
