+++ p1_key/test_5th_task.h
@@ -1,1 +1,1 @@
-bool test_5th_task(const bool, const unsigned short, const TS&, const bool, unsigned long int&, unsigned long int&);
+bool test_5th_task(const bool, const uint8_t, const TS&, const bool, unsigned long int&, unsigned long int&);
